+++ src/com/blo/sales/v2/controller/IUserController.java
@@ -8,5 +8,5 @@ public interface IUserController {
     
     PojoIntLoggedInUser doLogin(PojoIntUser userData) throws BloSalesV2Exception;
     
-    PojoIntUser getUserById(int id) throws BloSalesV2Exception;
+    PojoIntUser getUserById(long id) throws BloSalesV2Exception;
 }
