+++ src/com/blo/sales/v2/view/windows/pojos/PojoProduct.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public class PojoProduct {
     
-    private int idProduct;
+    private long idProduct;
     
     private String product;
     
@@ -18,13 +18,13 @@ public class PojoProduct {
     
     private String barCode;
     
-    private int fkCategory;
+    private long fkCategory;
 
-    public int getIdProduct() {
+    public long getIdProduct() {
         return idProduct;
     }
 
-    public void setIdProduct(int idProduct) {
+    public void setIdProduct(long idProduct) {
         this.idProduct = idProduct;
     }
 
@@ -76,11 +76,11 @@ public class PojoProduct {
         this.barCode = barCode;
     }
 
-    public int getFkCategory() {
+    public long getFkCategory() {
         return fkCategory;
     }
 
-    public void setFkCategory(int fkCategory) {
+    public void setFkCategory(long fkCategory) {
         this.fkCategory = fkCategory;
     }
